requirements.txt
Django>=4.2.4,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.2.0
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
